+++ services/continuous_querier/service_test.go
@@ -122,7 +122,7 @@ func TestContinuousQueryService_ResampleOptions(t *testing.T) {
 	s.QueryExecutor.StatementExecutor = &StatementExecutor{
 		ExecuteStatementFn: func(stmt influxql.Statement, ctx influxql.ExecutionContext) error {
 			s := stmt.(*influxql.SelectStatement)
-			min, max, err := influxql.TimeRange(s.Condition)
+			min, max, err := influxql.TimeRange(s.Condition, s.Location)
 			if err != nil {
 				t.Errorf("unexpected error parsing time range: %s", err)
 			} else if !expected.min.Equal(min) || !expected.max.Equal(max) {
@@ -203,7 +203,7 @@ func TestContinuousQueryService_EveryHigherThanInterval(t *testing.T) {
 	s.QueryExecutor.StatementExecutor = &StatementExecutor{
 		ExecuteStatementFn: func(stmt influxql.Statement, ctx influxql.ExecutionContext) error {
 			s := stmt.(*influxql.SelectStatement)
-			min, max, err := influxql.TimeRange(s.Condition)
+			min, max, err := influxql.TimeRange(s.Condition, s.Location)
 			if err != nil {
 				t.Errorf("unexpected error parsing time range: %s", err)
 			} else if !expected.min.Equal(min) || !expected.max.Equal(max) {
@@ -272,7 +272,7 @@ func TestContinuousQueryService_GroupByOffset(t *testing.T) {
 	s.QueryExecutor.StatementExecutor = &StatementExecutor{
 		ExecuteStatementFn: func(stmt influxql.Statement, ctx influxql.ExecutionContext) error {
 			s := stmt.(*influxql.SelectStatement)
-			min, max, err := influxql.TimeRange(s.Condition)
+			min, max, err := influxql.TimeRange(s.Condition, s.Location)
 			if err != nil {
 				t.Errorf("unexpected error parsing time range: %s", err)
 			} else if !expected.min.Equal(min) || !expected.max.Equal(max) {
@@ -432,7 +432,7 @@ func TestExecuteContinuousQuery_TimeRange(t *testing.T) {
 			s.QueryExecutor.StatementExecutor = &StatementExecutor{
 				ExecuteStatementFn: func(stmt influxql.Statement, ctx influxql.ExecutionContext) error {
 					s := stmt.(*influxql.SelectStatement)
-					min, max, err := influxql.TimeRange(s.Condition)
+					min, max, err := influxql.TimeRange(s.Condition, s.Location)
 					max = max.Add(time.Nanosecond)
 					if err != nil {
 						t.Errorf("unexpected error parsing time range: %s", err)
@@ -546,7 +546,7 @@ func TestExecuteContinuousQuery_TimeZone(t *testing.T) {
 				ExecuteStatementFn: func(stmt influxql.Statement, ctx influxql.ExecutionContext) error {
 					test := <-tests
 					s := stmt.(*influxql.SelectStatement)
-					min, max, err := influxql.TimeRange(s.Condition)
+					min, max, err := influxql.TimeRange(s.Condition, s.Location)
 					max = max.Add(time.Nanosecond)
 					if err != nil {
 						t.Errorf("unexpected error parsing time range: %s", err)